Remove opCast and castFrom
Those didn't work properly: If a LinearGradient was created like this:
Patter a = new Pattern(cairo_c_pointer);
this code would fail:
auto b = cast(Gradient)a;
(but this would work):
auto b = cast(LinearGradient)a;

It is now required that Patterns and Surfaces are created with the
correct D class from the start. Use either the correct class constructors
or the createFromNative methods.