Use C++11-compatible template parameters for Mersenne Twister
This patch reworks the existing Mersenne Twister template parameters to
match the names and types specified in the C++11 standard, replacing
`Uint` with `UIntType` and `uint` with `size_t`.

Besides bringing the design closer to the C++11 standard, these changes
also reduce ambiguity: `Uint` as a type name is visually easy to confuse
with `uint`, which can lead to typos and confusion about intentions.
`UIntType` may be more verbose, but it is much clearer in meaning and
intent.

Ordering and placement of the parameters in the specific 32- and 64-bit
instantiations of the template has been updated to match the standard.