Don't run meson build with Buildkite (#95)
Buildkite is used by DMD to ensure downstream projects are not broken.
Using meson here forces Buildkite agents to depend on meson.
While it would be possible to do so, that requires some work and coordination,
and might get delayed. In the meantime, just disable meson when buildkit is defined.