Correct version dependency on mir-core
Currently, programs depending on mir-random and mir-algorithm are broken
(e.g. lubeck) because of the version specification.
While it looks very much like a bug in dub, there doesn't seem to be a
tangible reason to have the current version restriction:
if v1.0.0 works, any v1.x.y should work as well.