Make tryMatch example compatible with noreturn inference (#49)
dlang/dmd#12553 will enable the frontend to infer `noreturn` as the
return type of templated / auto functions.

The unittest fails with this PR because `collideWith(Asteroid, Spaceship)`
always throws an exception and hence is inferred as `noreturn`. This
causes `collide(Asteroid, Spaceship)` to return `string` instead of
`void` because `noreturn` is implicitly convertible to any type.