Simplify deserialize on Variant!(void, T) (#28)
There was a special case for Nullable, I also applied this to Variants
with void in first place. This is used heavily in serve-d.

Additionally it seems Variants don't actually support the serdeEnumProxy
properly that's in the here adjusted test. However with Nullable and now
with Variant!(void, T) they at least work, which should cover a lot of
cases. In serve-d there is only a single exception to this, to which I
simply implemented a custom wrapper struct for serialization around.